pyproject.toml
[tool.poetry]
name = "drf-auth"
version = "0.1.0"
description = ""
authors = ["Yahialabeeb <yahia.labeeb@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djando = "^0.2"
djangorestframework = "^3.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
